done refactoring queue into separate files
